move setDatabase functionality to Jvminit and change to service name
